+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * When getting a registered member by an object, add same cv- property on the member as the object.
 * New feature: register member's pointer and (low-level) const pointer. ("register_member_ptr", "register_member_cptr")
 * New feature: register member's reference and (low-level) const reference. ("register_member_ref", "register_member_cref")
+* Use constexpr if statement if supports C++17.
 
 ## v1.1.2 - 2024.9.22